watcher/requirements.txt
watchdog
requests
haiku.rag
pypdf
ollama